+++ .isort.cfg
@@ -1,2 +1,2 @@
 [settings]
-known_third_party = dask,napari,numpy,ome_zarr,pytest,setuptools,vispy,zarr
+known_third_party = dask,napari,numpy,ome_zarr,pytest,setuptools,zarr
