+++ napari_ome_zarr/_tests/test_reader.py
@@ -8,10 +8,8 @@ from napari.utils.colormaps import AVAILABLE_COLORMAPS, Colormap
 from ome_zarr.data import astronaut, create_zarr
 from ome_zarr.writer import write_image, write_plate_metadata, write_well_metadata
 
-from napari_ome_zarr._reader import (
-    _match_colors_to_available_colormap,
-    napari_get_reader,
-)
+from napari_ome_zarr._reader import napari_get_reader
+from napari_ome_zarr.ome_zarr_reader import _match_colors_to_available_colormap
 
 
 class TestNapari:
